Add support for user activation
The activation is performed using a pre-generated token consisted of
hash on the identifier and user name.

Add boolean flag to the user account indicating that the user was
activated or not.

Update the login process to allow login of only active user accounts.

Add db migration to initialize the field.

Add rest endpoint for activation using token.

Refactor tests to construct active test user